Fix duplicate admin breadcrumbs on inventory detail pages
