+++ apps/app/app/admin/inventory/[inventoryId]/items/create/page.tsx
@@ -7,6 +7,7 @@ import { SelectItems } from '@signalco/ui-primitives/SelectItems';
 import { Stack } from '@signalco/ui-primitives/Stack';
 import { Typography } from '@signalco/ui-primitives/Typography';
 import { notFound } from 'next/navigation';
+import { AdminBreadcrumbLevelSelector } from '../../../../../../components/admin/navigation/AdminBreadcrumbLevelSelector';
 import { auth } from '../../../../../../lib/auth/auth';
 import { KnownPages } from '../../../../../../src/KnownPages';
 import { createInventoryItemAction } from '../../../../../(actions)/inventoryActions';
@@ -70,7 +71,10 @@ export default async function CreateInventoryItemPage({
         <Stack spacing={4}>
             <Breadcrumbs
                 items={[
-                    { label: 'Zalihe', href: KnownPages.Inventory },
+                    {
+                        label: <AdminBreadcrumbLevelSelector />,
+                        href: KnownPages.Inventory,
+                    },
                     {
                         label: config.label,
                         href: KnownPages.InventoryConfig(id),
