+++ apps/app/app/admin/inventory/[inventoryId]/page.tsx
@@ -20,6 +20,7 @@ import { Table } from '@signalco/ui-primitives/Table';
 import { Typography } from '@signalco/ui-primitives/Typography';
 import Link from 'next/link';
 import { notFound } from 'next/navigation';
+import { AdminBreadcrumbLevelSelector } from '../../../../../components/admin/navigation/AdminBreadcrumbLevelSelector';
 import { Field } from '../../../../components/shared/fields/Field';
 import { NoDataPlaceholder } from '../../../../components/shared/placeholders/NoDataPlaceholder';
 import { auth } from '../../../../lib/auth/auth';
@@ -53,7 +54,10 @@ export default async function InventoryConfigPage({
         <Stack spacing={2}>
             <Breadcrumbs
                 items={[
-                    { label: 'Zalihe', href: KnownPages.Inventory },
+                    {
+                        label: <AdminBreadcrumbLevelSelector />,
+                        href: KnownPages.Inventory,
+                    },
                     { label: config.label },
                 ]}
             />
