+++ apps/app/app/admin/inventory/create/page.tsx
@@ -6,6 +6,7 @@ import { Input } from '@signalco/ui-primitives/Input';
 import { SelectItems } from '@signalco/ui-primitives/SelectItems';
 import { Stack } from '@signalco/ui-primitives/Stack';
 import { Typography } from '@signalco/ui-primitives/Typography';
+import { AdminBreadcrumbLevelSelector } from '../../../../components/admin/navigation/AdminBreadcrumbLevelSelector';
 import { auth } from '../../../../lib/auth/auth';
 import { KnownPages } from '../../../../src/KnownPages';
 import { createInventoryConfigAction } from '../../../(actions)/inventoryActions';
@@ -25,7 +26,10 @@ export default async function CreateInventoryPage() {
         <Stack spacing={4}>
             <Breadcrumbs
                 items={[
-                    { label: 'Zalihe', href: KnownPages.Inventory },
+                    {
+                        label: <AdminBreadcrumbLevelSelector />,
+                        href: KnownPages.Inventory,
+                    },
                     { label: 'Nova zaliha' },
                 ]}
             />
